requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.4.26
click==8.2.0
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.78.1
passlib==1.7.4
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing_extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
